refactor: move server logic to lib
